Refactor and organize initialization functions
Move initialization and utility functions from minishell.c to new minishell_init.c for better modularity. Update Makefile and headers accordingly, and fix parser error status handling.
